move word_direction declaration to cw data types
